split into hello_world and counter
